+++ R/bigcurve-package.R
@@ -3,6 +3,7 @@
 
 ## usethis namespace: start
 #' @useDynLib bigcurve, .registration = TRUE
+#' @importFrom utils packageVersion
 ## usethis namespace: end
 NULL